Moved copy to own function and deleted dead code
