+++ pkg/binutils/constants.go
@@ -10,7 +10,6 @@ const (
 	gRPCGatewayEndpoint = ":8098"
 	gRPCDialTimeout     = 10 * time.Second
 
-	subnetEVMName        = "subnet-evm"
 	avalanchegoBinPrefix = "avalanchego-"
 	subnetEVMBinPrefix   = "subnet-evm-"
 	maxCopy              = 2147483648 // 2 GB
