+++ pkg/subnet/local_test.go
@@ -124,73 +124,6 @@ func TestDeployToLocal(t *testing.T) {
 	assert.Equal(testBlockChainID2, b.String())
 }
 
-func TestExistsWithLatestVersion(t *testing.T) {
-	assert := setupTest(t)
-
-	tmpDir := t.TempDir()
-	bc := binutils.NewBinaryChecker()
-
-	exists, latest, err := bc.ExistsWithLatestVersion(tmpDir, "avalanchego-v")
-	assert.NoError(err)
-	assert.False(exists)
-	assert.Empty(latest)
-
-	fake := filepath.Join(tmpDir, "anything")
-	err = os.Mkdir(fake, perms.ReadWriteExecute)
-	assert.NoError(err)
-	exists, latest, err = bc.ExistsWithLatestVersion(tmpDir, "avalanchego-v")
-	assert.NoError(err)
-	assert.False(exists)
-	assert.Empty(latest)
-
-	avagoOnly := filepath.Join(tmpDir, "avalanchego")
-	err = os.Mkdir(avagoOnly, perms.ReadWriteExecute)
-	assert.NoError(err)
-	exists, latest, err = bc.ExistsWithLatestVersion(tmpDir, "avalanchego-v")
-	assert.NoError(err)
-	assert.False(exists)
-	assert.Empty(latest)
-
-	existsOneOnly := filepath.Join(tmpDir, "avalanchego-v1.7.10")
-	err = os.Mkdir(existsOneOnly, perms.ReadWriteExecute)
-	assert.NoError(err)
-	exists, latest, err = bc.ExistsWithLatestVersion(tmpDir, "avalanchego-v")
-	assert.NoError(err)
-	assert.True(exists)
-	assert.Equal(existsOneOnly, latest)
-
-	ver1 := filepath.Join(tmpDir, "avalanchego-v1.8.0")
-	ver2 := filepath.Join(tmpDir, "avalanchego-v1.18.0")
-	ver3 := filepath.Join(tmpDir, "avalanchego-v1.8.1")
-	ver4 := filepath.Join(tmpDir, "avalanchego-v0.8.0")
-	ver5 := filepath.Join(tmpDir, "avalanchego-v0.88.0")
-	ver6 := filepath.Join(tmpDir, "avalanchego-v0.1.0")
-	ver7 := filepath.Join(tmpDir, "avalanchego-v0.11.0")
-	ver8 := filepath.Join(tmpDir, "avalanchego-0.11.0")
-
-	err = os.Mkdir(ver1, perms.ReadWriteExecute)
-	assert.NoError(err)
-	err = os.Mkdir(ver2, perms.ReadWriteExecute)
-	assert.NoError(err)
-	err = os.Mkdir(ver3, perms.ReadWriteExecute)
-	assert.NoError(err)
-	err = os.Mkdir(ver4, perms.ReadWriteExecute)
-	assert.NoError(err)
-	err = os.Mkdir(ver5, perms.ReadWriteExecute)
-	assert.NoError(err)
-	err = os.Mkdir(ver6, perms.ReadWriteExecute)
-	assert.NoError(err)
-	err = os.Mkdir(ver7, perms.ReadWriteExecute)
-	assert.NoError(err)
-	err = os.Mkdir(ver8, perms.ReadWriteExecute)
-	assert.NoError(err)
-
-	exists, latest, err = bc.ExistsWithLatestVersion(tmpDir, "avalanchego-v")
-	assert.NoError(err)
-	assert.True(exists)
-	assert.Equal(ver2, latest)
-}
-
 func TestGetLatestAvagoVersion(t *testing.T) {
 	assert := setupTest(t)
